v:1.01 hash function passing, rehashing, stability improvements
v:1.01 patch.

When creating hash table, now pass the hashing function as an input, this allows a new hash function to be passed when rehashing.

Added rehash function, which takes in a hash table, a new table size, and a new hash function, all elements from input hash function are copied to new hash table according to new table size and hashing function. Finally the original hash table is deallocated, and the pointer to it is reassigned to the new hash table.

Stability improvements: trying to search for an element in the table which doesn't exist no longer crashes, instead returns NULL.